pyproject.toml
[tool.poetry]
name = "simple-honcho-discord-bot"
version = "0.1.0"
description = "Simple Discord bot with Honcho storage backend"
authors = ["vintro <vince@plasticlabs.ai>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
py-cord = "^2.4.1"
python-dotenv = "^1.0.0"
langchain-core = "^0.1.12"
langchain-openai = "^0.0.2.post1"
honcho-ai = {path = "../../honcho/sdk", develop = true}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
